+++ src/charade/charade.tsx
@@ -8,26 +8,35 @@ import { COLOR } from '@core/colors/colors.constants'
 import { ThemeModel } from '@core/styled/models'
 import { CustomCounter } from '@core/custom-counter/custom-counter'
 import { useLocalize } from '@core/hooks/use-localize.hook'
-import { CharadeButtonsContainer, CharadeConfigurationContainer } from './charade.styled'
+import { CustomSwitch } from '@core/custom-switch/custom-switch'
 import { useCharadePlay } from './hooks'
+import { CharadeButtonsContainer, CharadeHeadline, CharadeSwitchWrapper, CharadeSeparator } from './charade.styled'
 
 export const Charade = () => {
   const localize = useLocalize()
   const theme = useTheme() as ThemeModel
   const { top: topInset } = useSafeAreaInsets()
-  const { handlePlayCharade, count, setCount } = useCharadePlay()
+  const { handlePlayCharade, count, setCount, allowDuplicatedLetters, setAllowDuplicatedLetters } = useCharadePlay()
 
   return (
     <SafeAreaFlexContainer backgroundColor={theme.backgroundPrimary}>
       <Header type="charade" />
-
-      <CharadeConfigurationContainer>
         <CustomCounter value={count} range={[ 3, 9 ]} setValue={setCount} />
 
         <CharadeButtonsContainer {...{ topInset }}>
           <CustomButton onPress={handlePlayCharade} color={COLOR.DARK_SEA_GREEN} title={localize().play} />
         </CharadeButtonsContainer>
-      </CharadeConfigurationContainer>
+
+        <CharadeSeparator />
+
+        <CharadeHeadline children={localize().allow_duplicated_letters} />
+        <CharadeSwitchWrapper>
+          <CustomSwitch
+            color={COLOR.DARK_SEA_GREEN}
+            defaultValue={allowDuplicatedLetters}
+            onValueChange={setAllowDuplicatedLetters}
+          />
+        </CharadeSwitchWrapper>
     </SafeAreaFlexContainer>
   )
 }