+++ src/charade/hooks/use-charade-play.hook.ts
@@ -8,6 +8,8 @@ interface UseCharadePlay {
   handlePlayCharade: () => void;
   count: number;
   setCount: (count: number) => void;
+  allowDuplicatedLetters: boolean;
+  setAllowDuplicatedLetters: (allowDuplicatedLetters: boolean) => void;
 }
 
 export const useCharadePlay = (): UseCharadePlay => {
@@ -15,13 +17,22 @@ export const useCharadePlay = (): UseCharadePlay => {
 
   const [ count, setCount ] = React.useState(DEFAULT_COUNT)
   const navigation = useNavigation<any>()
+  const [ allowDuplicatedLetters, setAllowDuplicatedLetters ] = React.useState(true)
 
   const handlePlayCharade = () => {
     const getWords = (words: string[]) => getRandomWords(words, (word: string) => word.length === count)
     const allWords = getWords(allWordsByLength)
-    const word = allWords[Math.floor(Math.random() * allWords.length)]?.toUpperCase?.()
+
+    let word = allWords[Math.floor(Math.random() * allWords.length)]?.toUpperCase?.()
+
+    if (!allowDuplicatedLetters) {
+      while([ ...new Set(word.split('')) ].length !== word.length) {
+        word = allWords[Math.floor(Math.random() * allWords.length)]?.toUpperCase?.()
+      }
+    }
+
     navigation.navigate(SCREEN.CHARADE_PLAY, { word, allWords })
   }
 
-  return { handlePlayCharade, count, setCount }
+  return { handlePlayCharade, count, setCount, allowDuplicatedLetters, setAllowDuplicatedLetters }
 }