+++ src/core/custom-switch/custom-switch.tsx
@@ -1,15 +1,15 @@
 import * as React from 'react'
 import { Switch } from 'react-native-paper'
-import { ViewProps } from 'react-native'
 import { COLOR } from '../colors/colors.constants'
 import { CustomSwitchContainer } from './custom-switch.styled'
 
 interface Props {
   defaultValue: boolean;
+  color?: COLOR;
   onValueChange: (value: boolean) => void;
 }
 
-export const CustomSwitch = ({ defaultValue, onValueChange }: Props) => {
+export const CustomSwitch = ({ defaultValue, color, onValueChange }: Props) => {
   const [ value, setValue ] = React.useState(defaultValue)
 
   const _onValueChange = React.useCallback((newValue: boolean) => {
@@ -19,7 +19,7 @@ export const CustomSwitch = ({ defaultValue, onValueChange }: Props) => {
 
   return (
     <CustomSwitchContainer>
-      <Switch color={COLOR.GOLD} value={value} onValueChange={_onValueChange} />
+      <Switch color={color ?? COLOR.GOLD} value={value} onValueChange={_onValueChange} />
     </CustomSwitchContainer>
   )
 }